requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.12.0
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.31
langchain-text-splitters==0.3.9
langsmith==0.4.15
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.2
openai==1.101.0
orjson==3.11.2
packaging==25.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.24.0
